start permission set

+++ routes/article.js
@@ -5,21 +5,31 @@ const router = require("express").Router();
 
 router.get("/", ensureAuthenticated, async (req, res) => {
 	const categories = await db.Categorie.findAll({ raw: true });
-	const articles = await db.Article.findAll({
-		include: ["Categorie"],
-		raw: true,
-	});
-	return res.render("articles/", { categories, articles });
+	if (req.user["Role.isAdmin"]) {
+		const articles = await db.Article.findAll({
+			include: ["Categorie"],
+			raw: true,
+		});
+		return res.render("articles/", { categories, articles });
+	} else {
+		const articles = await db.Article.findAll({
+			include: ["Categorie"],
+			raw: true,
+		});
+		return res.render("articles/", { categories, articles });
+	}
 });
 
 router.post("/add", ensureAuthenticated, (req, res, next) => {
-	db.Article.create(req.body)
-		.then(() => {
-			return res.redirect(req.headers.referer);
-		})
-		.catch((err) => {
-			next(err);
-		});
+	if (req.user["Role.isAdmin"]) {
+		db.Article.create(req.body)
+			.then(() => {
+				return res.redirect(req.headers.referer);
+			})
+			.catch((err) => {
+				next(err);
+			});
+	} else return res.status(401).send("Not allowed");
 });
 
 /*  edite  article*/
@@ -30,15 +40,17 @@ router.post(
 		const articleId = req.params.roleId; // Récupère l'ID du rôle à éditer
 		const updatedData = req.body; // Récupère les nouvelles données du rôle depuis le corps de la requête
 
-		db.Article.update(updatedData, {
-			where: { id: articleId },
-		})
-			.then(() => {
-				return res.redirect(req.headers.referer);
+		if (req.user["Role.isAdmin"]) {
+			db.Article.update(updatedData, {
+				where: { id: articleId },
 			})
-			.catch((err) => {
-				next(err);
-			});
+				.then(() => {
+					return res.redirect(req.headers.referer);
+				})
+				.catch((err) => {
+					next(err);
+				});
+		}
 	},
 );
 
@@ -49,15 +61,17 @@ router.get(
 	(req, res, next) => {
 		const articleId = req.params.roleId; // Récupère l'ID du rôle à supprimer
 
-		db.Article.destroy({
-			where: { id: articleId },
-		})
-			.then(() => {
-				return res.redirect(req.headers.referer);
+		if (req.user["Role.isAdmin"]) {
+			db.Article.destroy({
+				where: { id: articleId },
 			})
-			.catch((err) => {
-				next(err);
-			});
+				.then(() => {
+					return res.redirect(req.headers.referer);
+				})
+				.catch((err) => {
+					next(err);
+				});
+		}
 	},
 );
 module.exports = router;
